1.添加RateLimiter 2.添加一个不需要储存的NetworkResource，简化开发 3.登陆替换对应简单写法 4.添加了Room的使用例子

+++ common_ebd/src/main/java/cqebd/student/vo/User.kt
@@ -7,7 +7,7 @@ import com.google.gson.annotations.SerializedName
  * Created by @author xiaofu on 2018/12/18.
  */
 data class User(@SerializedName("studentId")
-                val ID: Long,
+                val ID: Int,
                 val Name: String,// 学生姓名
                 val LoginName: String,// 登陆账号
                 @SerializedName("Tel")
